Clean up RHEL/CentOS 7 compiler warnings.
git-svn-id: http://svn.eol.ucar.edu/svn/raf/trunk@7367 640d5228-2204-0410-8d71-8f46fa6850b4

+++ libraf++/adsIO.h
@@ -103,7 +103,7 @@ private:
   FILE	*fp;
 
   void	resetFile();
-  int	curr2dLR, currSyncLR;
+  size_t	curr2dLR, currSyncLR;
 
   char	physRecord[MX_PHYS];
   char	*syncPhysRecord, *twoDPhysRecord;
